src/backend/services/task/getter/requirements.txt
asn1crypto==1.5.1
greenlet==3.0.3
iniconfig==2.0.0
packaging==23.2
pg8000==1.30.4
pluggy==1.3.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0